thomas-android/src/main/java/com/servolabs/thomas/webservice/DateSerializer.java
package com.servolabs.thomas.webservice;

import com.google.gson.JsonElement;
import com.google.gson.JsonPrimitive;
import com.google.gson.JsonSerializationContext;
import com.google.gson.JsonSerializer;

import java.lang.reflect.Type;
import java.text.SimpleDateFormat;
import java.util.Date;

public class DateSerializer implements JsonSerializer<Date> {
    private static final SimpleDateFormat dateFormat = new SimpleDateFormat(RetrieveAvailableTrainingSessionsService.DATE_TEXT_FORMAT);

    @Override
    public JsonElement serialize(Date date, Type type, JsonSerializationContext jsonSerializationContext) {

        return date == null ? null : new JsonPrimitive(dateFormat.format(date));
    }
}
